rename spec-helpers.ts so it is not covered
