+++ server/src/model/bundle.spec.ts
@@ -1,6 +1,6 @@
 import expect from 'expect'
 import { Bundle, BundleValidationKind } from './bundle'
-import { bundleMaker, expectErrors, first, loadSuccess, makeBundle, read } from './spec-helpers'
+import { bundleMaker, expectErrors, first, loadSuccess, makeBundle, read } from './spec-helpers.spec'
 
 describe('Bundle validations', () => {
   it(BundleValidationKind.NO_BOOKS.title, () => {
