+++ server/src/model/fileish.spec.ts
@@ -1,7 +1,7 @@
 import expect from 'expect'
 import { Opt } from './utils'
 import { Fileish } from './fileish'
-import { first, FS_PATH_HELPER, makeBundle } from './spec-helpers'
+import { first, FS_PATH_HELPER, makeBundle } from './spec-helpers.spec'
 
 describe('The abstract ancestor class', () => {
   let previousNodeEnv: Opt<string>
